+++ lib/slop.rb
@@ -16,7 +16,7 @@ class Slop
   class InvalidOptionError < RuntimeError; end
 
   # @return [String] The current version string
-  VERSION = '1.5.5'
+  VERSION = '1.6.0'
 
   # Parses the items from a CLI format into a friendly object.
   #